bring in consistency checks
